Move exception logger to a dedicated class
